ref: Move files to the streaming_kafka_consumer package
Moves most of the files that will eventually form the streaming
consumer library into the streaming_kafka_consumer package.
Note the test files, streaming consumer factory and a few
other pieces have not yet been moved over.
