+++ snuba/cli/multistorage_consumer.py
@@ -4,6 +4,9 @@ from typing import Any, Optional, Sequence
 
 import click
 from confluent_kafka import Producer as ConfluentKafkaProducer
+from streaming_kafka_consumer.backends.kafka import KafkaConsumer
+from streaming_kafka_consumer.processing import StreamProcessor
+from streaming_kafka_consumer.types import Topic
 
 from snuba import environment, settings
 from snuba.consumers.consumer import MultistorageConsumerProcessingStrategyFactory
@@ -11,7 +14,6 @@ from snuba.datasets.storages import StorageKey
 from snuba.datasets.storages.factory import WRITABLE_STORAGES, get_writable_storage
 from snuba.environment import setup_logging, setup_sentry
 from snuba.utils.metrics.wrapper import MetricsWrapper
-from snuba.utils.streams.backends.kafka import KafkaConsumer
 from snuba.utils.streams.configuration_builder import (
     build_kafka_consumer_configuration,
     build_kafka_producer_configuration,
@@ -20,8 +22,6 @@ from snuba.utils.streams.kafka_consumer_with_commit_log import (
     KafkaConsumerWithCommitLog,
 )
 from snuba.utils.streams.metrics_adapter import StreamMetricsAdapter
-from snuba.utils.streams.processing import StreamProcessor
-from snuba.utils.streams.types import Topic
 
 logger = logging.getLogger(__name__)
