+++ snuba/consumers/consumer.py
@@ -20,6 +20,18 @@ from typing import (
 import rapidjson
 from confluent_kafka import Producer as ConfluentKafkaProducer
 from streaming_kafka_consumer import Message, Partition, Topic
+from streaming_kafka_consumer.backends.kafka import KafkaPayload
+from streaming_kafka_consumer.processing.strategies import ProcessingStrategy
+from streaming_kafka_consumer.processing.strategies import (
+    ProcessingStrategy as ProcessingStep,
+)
+from streaming_kafka_consumer.processing.strategies import ProcessingStrategyFactory
+from streaming_kafka_consumer.processing.strategies.streaming import (
+    CollectStep,
+    FilterStep,
+    ParallelTransformStep,
+    TransformStep,
+)
 
 from snuba.clickhouse.http import JSONRow, JSONRowEncoder
 from snuba.consumers.types import KafkaMessageMetadata
@@ -29,20 +41,8 @@ from snuba.datasets.storages import StorageKey
 from snuba.processor import InsertBatch, MessageProcessor, ReplacementBatch
 from snuba.utils.metrics import MetricsBackend
 from snuba.utils.metrics.wrapper import MetricsWrapper
-from snuba.utils.streams.backends.kafka import KafkaPayload
 from snuba.utils.streams.configuration_builder import build_kafka_producer_configuration
 from snuba.utils.streams.metrics_adapter import StreamMetricsAdapter
-from snuba.utils.streams.processing.strategies import ProcessingStrategy
-from snuba.utils.streams.processing.strategies import (
-    ProcessingStrategy as ProcessingStep,
-)
-from snuba.utils.streams.processing.strategies import ProcessingStrategyFactory
-from snuba.utils.streams.processing.strategies.streaming import (
-    CollectStep,
-    FilterStep,
-    ParallelTransformStep,
-    TransformStep,
-)
 from snuba.writer import BatchWriter
 
 logger = logging.getLogger("snuba.consumer")
