+++ snuba/consumers/strict_consumer.py
@@ -3,8 +3,7 @@ from enum import Enum
 from typing import Callable, MutableMapping, Optional, Sequence, Tuple
 
 from confluent_kafka import Consumer, KafkaError, Message, TopicPartition
-
-from snuba.utils.streams.backends.kafka.configuration import KafkaBrokerConfig
+from streaming_kafka_consumer.backends.kafka.configuration import KafkaBrokerConfig
 
 logger = logging.getLogger("snuba.kafka-consumer")
