+++ snuba/datasets/message_filters.py
@@ -1,8 +1,8 @@
 from abc import ABC, abstractmethod
 from typing import Generic
 
-from snuba.utils.streams.backends.kafka import KafkaPayload
-from snuba.utils.streams.types import Message, TPayload
+from streaming_kafka_consumer.backends.kafka import KafkaPayload
+from streaming_kafka_consumer.types import Message, TPayload
 
 
 class StreamMessageFilter(ABC, Generic[TPayload]):
