+++ snuba/replacer.py
@@ -5,14 +5,14 @@ from typing import Optional, Sequence
 
 import simplejson as json
 from streaming_kafka_consumer import Message
+from streaming_kafka_consumer.backends.kafka import KafkaPayload
 
 from snuba.clusters.cluster import ClickhouseClientSettings
 from snuba.datasets.storage import WritableTableStorage
 from snuba.processor import InvalidMessageVersion
 from snuba.replacers.replacer_processor import Replacement, ReplacementMessage
 from snuba.utils.metrics import MetricsBackend
-from snuba.utils.streams.backends.kafka import KafkaPayload
-from snuba.utils.streams.processing.strategies.batching import AbstractBatchWorker
+from snuba.utils.streams.batching import AbstractBatchWorker
 
 logger = logging.getLogger("snuba.replacer")
