--- snuba/utils/streams/backends/local/storages/__init__.py
@@ -1,5 +1,0 @@
-from snuba.utils.streams.backends.local.storages.abstract import MessageStorage
-
-__all__ = [
-    "MessageStorage",
-]
