--- snuba/utils/streams/processing/__init__.py
@@ -1,3 +1,0 @@
-from snuba.utils.streams.processing.processor import StreamProcessor
-
-__all__ = ["StreamProcessor"]
