+++ streaming_kafka_consumer/backends/local/storages/__init__.py
@@ -1,0 +1,5 @@
+from streaming_kafka_consumer.backends.local.storages.abstract import MessageStorage
+
+__all__ = [
+    "MessageStorage",
+]
