+++ streaming_kafka_consumer/processing/__init__.py
@@ -1,0 +1,3 @@
+from streaming_kafka_consumer.processing.processor import StreamProcessor
+
+__all__ = ["StreamProcessor"]
