+++ streaming_kafka_consumer/processing/strategies/streaming/transform.py
@@ -21,12 +21,12 @@ from typing import (
     TypeVar,
 )
 
-from snuba.utils.streams.metrics import DummyMetricsBackend, Gauge, Metrics
-from snuba.utils.streams.processing.strategies.abstract import MessageRejected
-from snuba.utils.streams.processing.strategies.abstract import (
+from streaming_kafka_consumer.metrics import DummyMetricsBackend, Gauge, Metrics
+from streaming_kafka_consumer.processing.strategies.abstract import MessageRejected
+from streaming_kafka_consumer.processing.strategies.abstract import (
     ProcessingStrategy as ProcessingStep,
 )
-from snuba.utils.streams.types import Message, TPayload
+from streaming_kafka_consumer.types import Message, TPayload
 
 logger = logging.getLogger(__name__)
