+++ tests/datasets/cdc/test_message_filters.py
@@ -1,9 +1,9 @@
 from datetime import datetime
 
 from streaming_kafka_consumer import Message, Partition, Topic
+from streaming_kafka_consumer.backends.kafka import KafkaPayload
 
 from snuba.datasets.cdc.message_filters import CdcTableNameMessageFilter
-from snuba.utils.streams.backends.kafka import KafkaPayload
 
 
 def test_table_name_filter() -> None:
