+++ tests/datasets/test_errors_replacer.py
@@ -9,6 +9,7 @@ import pytest
 import pytz
 import simplejson as json
 from streaming_kafka_consumer import Message, Partition, Topic
+from streaming_kafka_consumer.backends.kafka import KafkaPayload
 
 from snuba import replacer, settings
 from snuba.clickhouse import DATETIME_FORMAT
@@ -20,7 +21,6 @@ from snuba.optimize import run_optimize
 from snuba.redis import redis_client
 from snuba.settings import PAYLOAD_DATETIME_FORMAT
 from snuba.utils.metrics.backends.dummy import DummyMetricsBackend
-from snuba.utils.streams.backends.kafka import KafkaPayload
 from tests.fixtures import get_raw_event
 from tests.helpers import write_unprocessed_events
