+++ tests/test_replacer.py
@@ -6,6 +6,7 @@ from functools import partial
 import pytz
 import simplejson as json
 from streaming_kafka_consumer import Message, Partition, Topic
+from streaming_kafka_consumer.backends.kafka import KafkaPayload
 
 from snuba import replacer, settings
 from snuba.clickhouse import DATETIME_FORMAT
@@ -18,7 +19,6 @@ from snuba.optimize import run_optimize
 from snuba.redis import redis_client
 from snuba.settings import PAYLOAD_DATETIME_FORMAT
 from snuba.utils.metrics.backends.dummy import DummyMetricsBackend
-from snuba.utils.streams.backends.kafka import KafkaPayload
 from tests.fixtures import get_raw_event
 from tests.helpers import write_unprocessed_events
