+++ tests/utils/streams/backends/test_kafka.py
@@ -9,20 +9,20 @@ from unittest import TestCase
 
 import pytest
 from confluent_kafka.admin import AdminClient, NewTopic
-
-from snuba.utils.streams.backends.abstract import ConsumerError, EndOfPartition
-from snuba.utils.streams.backends.kafka import (
+from streaming_kafka_consumer.backends.abstract import ConsumerError, EndOfPartition
+from streaming_kafka_consumer.backends.kafka import (
     KafkaConsumer,
     KafkaPayload,
     KafkaProducer,
     as_kafka_configuration_bool,
 )
+from streaming_kafka_consumer.synchronized import Commit, commit_codec
+from streaming_kafka_consumer.types import Message, Partition, Topic
+
 from snuba.utils.streams.configuration_builder import get_default_kafka_configuration
 from snuba.utils.streams.kafka_consumer_with_commit_log import (
     KafkaConsumerWithCommitLog,
 )
-from snuba.utils.streams.synchronized import Commit, commit_codec
-from snuba.utils.streams.types import Message, Partition, Topic
 from tests.backends.confluent_kafka import FakeConfluentKafkaProducer
 from tests.utils.streams.backends.mixins import StreamsTestMixin
