+++ tests/utils/streams/processing/strategies/test_batching.py
@@ -3,16 +3,17 @@ from datetime import datetime
 from typing import Any, MutableSequence, Sequence
 from unittest.mock import patch
 
+from streaming_kafka_consumer.backends.local.backend import LocalBroker as Broker
+from streaming_kafka_consumer.backends.local.backend import LocalConsumer
+from streaming_kafka_consumer.processing.processor import StreamProcessor
+from streaming_kafka_consumer.types import Message, Topic
+
 from snuba.utils.metrics.backends.dummy import DummyMetricsBackend
-from snuba.utils.streams.backends.local.backend import LocalBroker as Broker
-from snuba.utils.streams.backends.local.backend import LocalConsumer
-from snuba.utils.streams.metrics_adapter import StreamMetricsAdapter
-from snuba.utils.streams.processing.processor import StreamProcessor
-from snuba.utils.streams.processing.strategies.batching import (
+from snuba.utils.streams.batching import (
     AbstractBatchWorker,
     BatchProcessingStrategyFactory,
 )
-from snuba.utils.streams.types import Message, Topic
+from snuba.utils.streams.metrics_adapter import StreamMetricsAdapter
 
 
 class FakeWorker(AbstractBatchWorker[int, int]):
